change serial to double

+++ Assignment_3/assignment3/serialConvolution.cu
@@ -37,10 +37,10 @@ const int mask[3][3] = {
 // };
 
 
-void printArray(int **array, int r, int c) {
+void printArray(double **array, int r, int c) {
     for (int i = 0; i < r; i++) {
         for (int j = 0; j < c; j++) {
-            printf("%d ", array[i][j]);
+            printf("%3.6f ", array[i][j]);
         }
         printf("\n");
     }
@@ -51,16 +51,16 @@ int randomNumberGeneration(int upperBound, int lowerBound) {
     return num;
 }
 
-int **createMatrix(int m, int n) {
-    int **array = (int **)malloc(m * sizeof(int *));
+double **createMatrix(int m, int n) {
+    double **array = (double **)malloc(m * sizeof(double *));
     for (int i = 0; i < m; i++) {
-        array[i] = (int *)malloc(n * sizeof(int));
+        array[i] = (double *)malloc(n * sizeof(double));
     }
 
     return array;
 }
 
-int **createData(int **array, int size, int dimension) {
+double **createData(double **array, int size, int dimension) {
     for (int i = 0; i < size; i++) {
         for (int j = 0; j < dimension; j++) {
             array[i][j] = randomNumberGeneration(9, 1);
@@ -69,7 +69,7 @@ int **createData(int **array, int size, int dimension) {
     return array;
 }
 
-int **padArray(int **input, int **output) {
+double **padArray(double **input, double **output) {
     int range = output_size - offset;
     // printf("%d \n", range);
 
@@ -82,7 +82,7 @@ int **padArray(int **input, int **output) {
     return output;
 }
 
-int **unpad(int **input, int **output) {
+double **unpad(double **input, double **output) {
     int range = output_size - offset;
 
     // unpad the array
@@ -95,11 +95,11 @@ int **unpad(int **input, int **output) {
 }
 
 
-int applyMask(int **array, int row, int col){
-    int n_size = offset * 2 + 1;
+double applyMask(double **array, int row, int col){
+    double n_size = offset * 2 + 1;
 
     // neighbours of giving location
-    int **neighbours = createMatrix(n_size, n_size);
+    double **neighbours = createMatrix(n_size, n_size);
 
     // int range = output_size - offset;
     // for (int i=row; i < range; i++){
@@ -122,8 +122,8 @@ int applyMask(int **array, int row, int col){
     neighbours[2][2] = array[row+1][col+1]; //bottom_right
 
 
-    int **convolution = createMatrix(n_size, n_size);
-    int value = 0;
+    double **convolution = createMatrix(n_size, n_size);
+    double value = 0;
 
     for (int r=0; r<3; r++){
         for(int c=0; c<3; c++){
@@ -137,7 +137,7 @@ int applyMask(int **array, int row, int col){
     return value;
 }
 
-int **serial_convolution(int **input, int **output){
+double **serial_convolution(double **input, double **output){
     int range = output_size - offset;
     // printf("%d ", range);
 
@@ -151,10 +151,10 @@ int **serial_convolution(int **input, int **output){
 
 
 int main(int argc, char **argv){
-    int **input = createMatrix(size,size);
-    int **padded = createMatrix(output_size, output_size);
-    int **output = createMatrix(output_size, output_size);
-    int **unpadded = createMatrix(output_size, output_size);
+    double **input = createMatrix(size,size);
+    double **padded = createMatrix(output_size, output_size);
+    double **output = createMatrix(output_size, output_size);
+    double **unpadded = createMatrix(output_size, output_size);
 
     input = createData(input, size, size);
     // printArray(input, size, size);
